requirements.txt
numpy==1.25.1
opencv_python==4.8.0.74
Pillow==10.0.0
Pillow==10.0.0
PyAutoGUI==0.9.54
pynput==1.7.6
PyScreeze==0.1.29
pipreqs==0.4.13
